Disabled the retry queue. Added config setting for retry queue. Events are saved when recieved and given an internal id. Added new internal class for events, replacing publishRequest. This will keep the internal id. Added Add/Get/List for the above event type and unit tests. Added IsContentMatch() to enable filtering events on content as well as topic and unit test. Added Subscribe(Subscription) method to enable subscription by content. (persistent content filter is not ready, see below) Added new properties to Subscription to diagnose missed events.
Work in progress (started but not ready):
The retry queue will use a database instead of memory. Failed events for a subscription will just use the reference ID and not store content.
The List<LinkSmartEvent> metod will be used to enable replay of events for a specific time period and (supplied) subscription: Replay(subscription, startTime, endTime)
Known problem: Persistent subscriptions - the content filter is not stored correctly.


git-svn-id: http://svn.code.sf.net/p/linksmart/code/trunk@950 95b9cf05-cb7e-4607-9643-a8f9ddef502b